Use the current interpreter's epylint.lint
epylint.py_run was calling *epylint* script in order to
do the computation. But this was prone to failures,
in the case where we have multiple installations of Pylint,
for different Python versions, which actually led to crashes
when a Python 2's py_run call was calling epylint's belonging
to Python 3. Instead of calling epylint script directly,
we are calling Python with our known epylint code.
In order to do this, we also had to switch the shell argument
to False.

Closes #441